// src/db/database.types.ts
export interface Database {
  graphql_public: {
    Tables: Record<never, never>;
    Views: Record<never, never>;
    Functions: {
      graphql: {
        Args: {
          extensions?: Json;
          operationName?: string;
          query?: string;
          variables?: Json;
        };
        Returns: Json;
      };
    };
    Enums: Record<never, never>;
    CompositeTypes: Record<never, never>;
  };
  public: {
    Tables: {
      ski_spec_notes: {
        Row: {
          content: string;
          created_at: string;
          id: string;
          ski_spec_id: string;
          updated_at: string;
        };
        Insert: {
          content: string;
          created_at?: string;
          id?: string;
          ski_spec_id: string;
          updated_at?: string;
        };
        Update: {
          content?: string;
          created_at?: string;
          id?: string;
          ski_spec_id?: string;
          updated_at?: string;
        };
        Relationships: [
          {
            foreignKeyName: 'ski_spec_notes_ski_spec_id_fkey';
            columns: ['ski_spec_id'];
            isOneToOne: false;
            referencedRelation: 'ski_specs';
            referencedColumns: ['id'];
          },
        ];
      };
      ski_specs: {
        Row: {
          algorithm_version: string;
          created_at: string;
          description: string | null;
          id: string;
          length: number;
          name: string;
          radius: number;
          relative_weight: number;
          surface_area: number;
          tail: number;
          tip: number;
          updated_at: string;
          user_id: string;
          waist: number;
          weight: number;
        };
        Insert: {
          algorithm_version: string;
          created_at?: string;
          description?: string | null;
          id?: string;
          length: number;
          name: string;
          radius: number;
          relative_weight: number;
          surface_area: number;
          tail: number;
          tip: number;
          updated_at?: string;
          user_id: string;
          waist: number;
          weight: number;
        };
        Update: {
          algorithm_version?: string;
          created_at?: string;
          description?: string | null;
          id?: string;
          length?: number;
          name?: string;
          radius?: number;
          relative_weight?: number;
          surface_area?: number;
          tail?: number;
          tip?: number;
          updated_at?: string;
          user_id?: string;
          waist?: number;
          weight?: number;
        };
        Relationships: [];
      };
    };
    Views: Record<never, never>;
    Functions: Record<never, never>;
    Enums: Record<never, never>;
    CompositeTypes: Record<never, never>;
  };
}

type DatabaseWithoutInternals = Omit<Database, '__InternalSupabase'>;

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, 'public'>];
// ...
export type TablesInsert<
  DefaultSchemaTableNameOrOptions extends keyof DefaultSchema['Tables'] | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions['schema']]['Tables']
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions['schema']]['Tables'][TableName] extends {
      Insert: infer I;
    }
    ? I
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema['Tables']
    ? DefaultSchema['Tables'][DefaultSchemaTableNameOrOptions] extends {
        Insert: infer I;
      }
      ? I
      : never
    : never;

export type TablesUpdate<
  DefaultSchemaTableNameOrOptions extends keyof DefaultSchema['Tables'] | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions['schema']]['Tables']
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions['schema']]['Tables'][TableName] extends {
      Update: infer U;
    }
    ? U
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema['Tables']
    ? DefaultSchema['Tables'][DefaultSchemaTableNameOrOptions] extends {
        Update: infer U;
      }
      ? U
      : never
    : never;